pyproject.toml
[tool.poetry]
name = "daily-report-bot"
version = "0.1.0"
description = ""
authors = ["antowkaka <dao9714@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.0"
python-dotenv = "^1.0.1"
motor = "^3.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
